requirements.txt
requests==2.32.3
pypdf==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dbt-core 
dbt-postgres